Add support for verify-balance operations.
This allows lines with the form:
 verify-balance <date> <account> <amount>
to be added to the input file.

When a file containing lines like this are read, the software will
check that the relevant accounts had the relevant balances on the
specified dates. If any of the assertions turn out to be incorrect,
the software will default to exiting after complaining about the error.

Relevant options:
* --ignore-balance-verification-failure
   If this option is specified, a failed balance verification will not
   make the program exit. This will allow any specified reports on the
   file to be generated. (Useful, e.g. if you want to use --print-register
   to try to compare the incorrect balances in an account with the true
   balances in a bank statement.)

* --verbose, or
  --show-balance-verifications
   If either of these are specified, a line confirming the success of each
   verify-balance will be printed to standard output.