New option --ignore-transactions-outside-dates.
Optionally ignore transactions occuring before --first-date or after
--last-date.  This is mainly useful with the --print-register command
that shows an account's running balance between two dates.

If this option *is* used, any transactions before the --first-date will be
ignored and the running balance reported will therefore start at zero.
If this is option is *not* used, any transactions occuring before --first-date
*will* affect the account's balance even though those transactions are not
shown, so the balance history shown might not start at zero.

Probably not worthwhile, but I couldn't decide which way the report should
work, so made it an option.